+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ## 0.14.19 - 2026-06-23
 * Require Python 3.11-3.14
 * Replace pyOpenSSL with the cryptography library
+* Require Docker Compose V2 (`docker compose`)
 
 ## 0.14.15 -2022-10-06
 * Updated pyOpenssl